+++ performance-tools/src/test/java/com/fillumina/performance/producer/progression/IterationProgressionPerformanceInstrumenterTest.java
@@ -33,7 +33,7 @@ public class IterationProgressionPerformanceInstrumenterTest {
             return this;
         }
 
-        private void assertSamples(final int... iterations) {
+        private void assertSamples(final long... iterations) {
             final AssertIterationsPerformanceConsumer assertIterations =
                     new AssertIterationsPerformanceConsumer()
                         .setIterations(iterations)
